agents/project_management_agent/requirements.txt
google-adk
python-dotenv
requests
beautifulsoup4
pydantic
uvicorn
fastapi
